ContributionCardV2 cleanup
